backend/requirements.txt
fastapi 
uvicorn
python-keycloak
requests
pydantic
python-multipart
sqlalchemy
arrow
psycopg
pydantic-settings
python-codicefiscale
sqlmodel #not being used currently, but would fix a the TicketList problem
opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy